Integrate wasmcli into wasmd
