+++ .gitignore
@@ -22,7 +22,6 @@ docs/node_modules
 # Data - ideally these don't exist
 baseapp/data/*
 client/lcd/keys/*
-cmd/wasmcli/statik/statik.go
 mytestnet
 
 # Testing
