--- docker/logs.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-tail -f /root/log/wasmd.log
