--- docker/run_rest_server.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-#set -euo pipefail
-
-wasmcli rest-server --laddr tcp://0.0.0.0:1317 --trust-node --unsafe-cors
