+++ docker/run_wasmd.sh
@@ -3,7 +3,6 @@
 if test -n "$1"; then
     # need -R not -r to copy hidden files
     cp -R "$1/.wasmd" /root
-    cp -R "$1/.wasmcli" /root
 fi
 
 mkdir -p /root/log
