+++ x/wasm/Governance.md
@@ -76,7 +76,7 @@ As gov proposals bypass the existing authorzation policy they are diabled and re
 ## CLI
 
 ```shell script
-  wasmcli tx gov submit-proposal [command]
+  wasmd tx gov submit-proposal [command]
 
 Available Commands:
   wasm-store           Submit a wasm binary proposal
